Refactor in preparation for IPC
